backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.0
gunicorn>=21.2
